기능 수정, 기능 추가
- 기간트 스킬 버그 수정
*움직이는 성채: 자신을 대상으로 지정할 수 있던 버그 수정
*쇄국: 아군뿐만 아니라 적군 유닛에게도 중첩 버프가 적용되던 버그 수정
- 스탯, 스킬 계산식 수정
*적의 수치는 원래값 그대로 사용하게 수정
*6개 스탯 수치도 원래값 그대로 사용하게 수정
* 스킬의 최종 대미지는 소수점 버림으로 계산하게 수정
*Hp, Mp 증가 및 감소할 때 소수점 버림으로 계산되게 수정

- 휴식, 진정 액션 기능 추가
*휴식: 최대HP의 일정 수치만큼 회복 후 턴 종료
*진정: 최대MP의 일정 수치만큼 회복, 현재 Rage값의 일정 수치만큼 감소 후 턴 종료

+++ Assets/Resources/Scripts/Battle/Skill/Player/Active/ParametricHealSkill.cs
@@ -46,7 +46,7 @@ public class ParametricHealSkill : SkillAsset, ITargetMapProvider
     int CalcHealAmount(BattleUnit caster, BattleUnit target)
     {
         // 마법공격력 * 배수. 필요하면 라우터/상태/장비에 따른 보정도 추가 가능
-        int baseStat = Mathf.Max(1, caster.MagicDamage);
+        float baseStat = Mathf.Max(1, caster.MagicDamage);
         float mult = Mathf.Max(0f, power);
         return Mathf.Max(1, Mathf.FloorToInt(baseStat * mult));
     }
